+++ client/Packages/com.beamable/Editor/BeamEditor.cs
@@ -20,6 +20,7 @@ using Beamable.Editor.Modules.EditorConfig;
 using Beamable.Editor.Realms;
 using Beamable.Editor.Reflection;
 using Beamable.Editor.ToolbarExtender;
+using Beamable.Editor.Toolbox.Models;
 using Beamable.Editor.UI;
 using Beamable.Inventory.Scripts;
 using Beamable.Reflection;
@@ -147,7 +148,7 @@ namespace Beamable
 					{
 						if (allowRetry)
 						{
-							BeamEditorContext.WriteConfig("", "");
+							BeamEditorContext.WriteConfig(string.Empty, string.Empty);
 							return TryInitConfigDatabase(false);
 						}
 						else
@@ -277,6 +278,8 @@ namespace Beamable
 			BeamEditorContextDependencies.AddSingleton(_ => HintPreferencesManager);
 			BeamEditorContextDependencies.AddSingleton<BeamableVsp>();
 
+			BeamEditorContextDependencies.AddSingleton<IToolboxViewService, ToolboxViewService>();
+
 			var hintReflectionSystem = GetReflectionSystem<BeamHintReflectionCache.Registry>();
 			foreach (var globallyAccessibleHintSystem in hintReflectionSystem.GloballyAccessibleHintSystems)
 				BeamEditorContextDependencies.AddSingleton(globallyAccessibleHintSystem.GetType(), () => globallyAccessibleHintSystem);
@@ -288,9 +291,11 @@ namespace Beamable
 			{
 				await BeamEditorContext.Default.InitializePromise;
 
+#if BEAMABLE_DEVELOPER
 				Debug.Log($"Initialized Default Editor Context [{BeamEditorContext.Default.PlayerCode}] - " +
 						  $"[{BeamEditorContext.Default.ServiceScope.GetService<PlatformRequester>().Cid}] - " +
 						  $"[{BeamEditorContext.Default.ServiceScope.GetService<PlatformRequester>().Pid}]");
+#endif
 				IsInitialized = true;
 
 				// Initialize toolbar
@@ -441,7 +446,7 @@ namespace Beamable
 
 			if (!ConfigFileExists)
 			{
-				SaveConfig("", "", BeamableEnvironment.ApiUrl);
+				SaveConfig(string.Empty, string.Empty, BeamableEnvironment.ApiUrl);
 				Logout();
 				InitializePromise = Promise.Success;
 				return;
@@ -457,7 +462,7 @@ namespace Beamable
 
 			if (string.IsNullOrEmpty(cid)) // with no cid, we cannot be logged in.
 			{
-				SaveConfig("", "", BeamableEnvironment.ApiUrl);
+				SaveConfig(string.Empty, string.Empty, BeamableEnvironment.ApiUrl);
 				Logout();
 				InitializePromise = Promise.Success;
 				return;
@@ -729,7 +734,7 @@ namespace Beamable
 				containerPrefix = containerPrefix
 			};
 
-			string path = "Assets/Beamable/Resources/config-defaults.txt";
+			string path = ConfigDatabase.GetFullPath("config-defaults");
 			var asJson = JsonUtility.ToJson(config, true);
 
 			var writeConfig = true;
@@ -744,7 +749,11 @@ namespace Beamable
 
 			if (writeConfig)
 			{
-				Directory.CreateDirectory("Assets/Beamable/Resources/");
+				string directoryName = Path.GetDirectoryName(path);
+				if (!string.IsNullOrWhiteSpace(directoryName))
+				{
+					Directory.CreateDirectory(directoryName);
+				}
 
 				if (File.Exists(path))
 				{
@@ -928,7 +937,6 @@ namespace Beamable
 				throw new Exception("Cannot switch to a realm with a null pid");
 			}
 
-			await ServiceScope.GetService<ContentIO>().FetchManifest();
 			var realms = await ServiceScope.GetService<RealmsService>().GetRealms(game);
 			var set = EditorPrefHelper
 					  .GetMap(REALM_PREFERENCE)
@@ -939,10 +947,23 @@ namespace Beamable
 			if (CurrentRealm == null || !CurrentRealm.Equals(realm))
 			{
 				CurrentRealm = realm;
+				await SaveRealmInConfig();
+				await ServiceScope.GetService<ContentIO>().FetchManifest();
 				OnRealmChange?.Invoke(realm);
+				ProductionRealm = game;
+				return;
 			}
-			ProductionRealm = game;
+			else
+			{
+				await ServiceScope.GetService<ContentIO>().FetchManifest();
+			}
 
+			ProductionRealm = game;
+			await SaveRealmInConfig();
+		}
+
+		private async Promise SaveRealmInConfig()
+		{
 			// Ensure we save the current cached data for domain reloads.
 			await SaveLastAuthenticatedUserDataForToken(Requester.Token, CurrentUser, CurrentCustomer, CurrentRealm);
 			SaveConfig(CurrentCustomer.Alias, CurrentRealm.Pid, cid: CurrentCustomer.Cid);