+++ client/Packages/com.beamable/Editor/UI/Common/UIElementsPolyfill.cs
@@ -1,3 +1,4 @@
+using Beamable.Common;
 using Beamable.Editor.UI.Components;
 using System;
 using System.Collections.Generic;
@@ -512,7 +513,7 @@ public static class UssLoader
 
 		if (u2021Available)
 		{
-#if UNITY_2021_2_OR_NEWER // 2021 is the max supported version, so we forward lean and assume all uss works in 2022.
+#if UNITY_2021_1_OR_NEWER // 2021 is the max supported version, so we forward lean and assume all uss works in 2022.
 			ussPaths.Add(u2021Path);
 #endif
 		}