+++ microservice/microserviceTests/microservice/dbmicroservice/BeamableMicroServiceTests/MethodSerializationTests.cs
@@ -46,7 +46,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "Sum", 1, 0, null));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "Sum", 1, 1, null));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -77,7 +77,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "Sum", 1, 0, new int[]{}));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "Sum", 1, 1, new int[]{}));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -107,7 +107,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "Sum", 1, 0, new int[]{1,2,3}));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "Sum", 1, 1, new int[]{1,2,3}));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -137,7 +137,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 0, null, null));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 1, null, null));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -167,7 +167,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 0, null, new int[]{1,2}));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 1, null, new int[]{1,2}));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -197,7 +197,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 0, new int[]{1,2}, null));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 1, new int[]{1,2}, null));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -227,7 +227,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 0, new int[]{1,2}, new int[]{3,4}));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "TwoArrays", 1, 1, new int[]{1,2}, new int[]{3,4}));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -257,7 +257,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "PromiseTestMethod", 1, 0));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "PromiseTestMethod", 1, 1));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -286,7 +286,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "PromiseTypelessTestMethod", 1, 0));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "PromiseTypelessTestMethod", 1, 1));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -331,7 +331,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          Assert.IsTrue(ms.HasInitialized);
 
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithJSON_AsParameter", 1, 0, serialized));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithJSON_AsParameter", 1, 1, serialized));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -362,7 +362,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithRegularString_AsParameter", 1, 0, "test_String"));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithRegularString_AsParameter", 1, 1, "test_String"));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -395,7 +395,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
 
          // testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithRegularString_AsParameter", 1, 0, "test_String"));
 
-         testSocket.SendToClient(ClientRequest.ClientCallablePayloadArgs("micro_sample", "MethodWithRegularString_AsParameter", 1, 0, "[\"test_String\"]"));
+         testSocket.SendToClient(ClientRequest.ClientCallablePayloadArgs("micro_sample", "MethodWithRegularString_AsParameter", 1, 1, "[\"test_String\"]"));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
@@ -426,83 +426,13 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithVector2Int_AsParameter", 1, 0, new Vector2Int(10, 20)));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithVector2Int_AsParameter", 1, 1, new Vector2Int(10, 20)));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);
          Assert.IsTrue(testSocket.AllMocksCalled());
       }
-      
-      [Test]
-      [NonParallelizable]
-      public async Task Call_LeaderboardCreate_FromTemplate()
-      {
-         LoggingUtil.Init();
-         TestSocket testSocket = null;
-         var version = "623";
-         var contentId = "leaderboards.test";
 
-         var leaderboardTemplate =
-            $"{{\"id\":\"{contentId}\",\"version\":\"{version}\",\"properties\":{{\"permissions\":{{\"data\":{{\"write_self\":true}}}}}}}}";
-         var expectedLeaderboardBody = "{\"permissions\": { \"write_self\": true}}";
-         var expectedJObject = JObject.Parse(expectedLeaderboardBody);
-         var ms = new BeamableMicroService(new TestSocketProvider(socket =>
-         {
-            testSocket = socket;
-            socket.AddStandardMessageHandlers()
-               .AddInitialContentMessageHandler(-5, new ContentReference
-               {
-                  id = contentId,
-                  version = version,
-                  visibility = Constants.Features.Content.PUBLIC
-               })
-               .AddMessageHandler(
-                  MessageMatcher
-                     .WithBody<JObject>(n => JToken.DeepEquals(expectedJObject, n)),
-                  MessageResponder.NoResponse(),
-                  MessageFrequency.OnlyOnce());
-         }), new TestContentResolver(x => Task.FromResult(leaderboardTemplate)));
-
-         await ms.Start<SimpleMicroservice>(new TestArgs());
-         Assert.IsTrue(ms.HasInitialized);
-         
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "LeaderboardCreateFromTemplateCallableTest", 1, 0, "leaderboard", contentId));
-
-         // simulate shutdown event...
-         await ms.OnShutdown(this, null);
-         Assert.IsTrue(testSocket.AllMocksCalled());
-      }
-      
-      [Test]
-      [NonParallelizable]
-      public async Task Call_LeaderboardCreate_FromCode()
-      {
-         LoggingUtil.Init();
-         TestSocket testSocket = null;
-
-         var expectedLeaderboardBody = "{\"permissions\": { \"write_self\": true}}";
-         var expectedJObject = JObject.Parse(expectedLeaderboardBody);
-         var ms = new BeamableMicroService(new TestSocketProvider(socket =>
-         {
-            testSocket = socket;
-            socket.AddStandardMessageHandlers()
-               .AddMessageHandler(
-                  MessageMatcher
-                     .WithBody<JObject>(n => JToken.DeepEquals(expectedJObject, n)),
-                  MessageResponder.NoResponse(),
-                  MessageFrequency.OnlyOnce());
-         }));
-
-         await ms.Start<SimpleMicroservice>(new TestArgs());
-         Assert.IsTrue(ms.HasInitialized);
-         
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "LeaderboardCreateFromCodeCallableTest", 1, 0, "leaderboard"));
-
-         // simulate shutdown event...
-         await ms.OnShutdown(this, null);
-         Assert.IsTrue(testSocket.AllMocksCalled());
-      }
-      
       [Test]
       [NonParallelizable]
       public async Task Call_MethodWithException()
@@ -526,7 +456,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
          await ms.Start<SimpleMicroservice>(new TestArgs());
          Assert.IsTrue(ms.HasInitialized);
 
-         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithExceptionThrow", 1, 0, string.Empty));
+         testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "MethodWithExceptionThrow", 1, 1, string.Empty));
 
          // simulate shutdown event...
          await ms.OnShutdown(this, null);